removes properties in inherited class
